feat(ui): unify forging assignment on the mobile design + receive redesign + send phone compaction (Android review round 5)
Forging assignment (Option A, mobile as the base per Johnny): ONE
ForgingAssignmentComponent for /forging-assignment and /wallet/assignment.
The desktop's Core-mode branches are ported in (node get_assignment/create/
revoke RPCs, per-preset estimatesmartfee, Core address list, wallet unlock
via PassphraseDialog — the mobile page gains unlock handling it never had);
remote keeps the client-side btcx path. Desktop rendering: 600px column with
send-style uppercase section labels above the inputs; phone keeps the exact
mobile look. Old 3-tab desktop page and free-text plot-address entry are
retired; WalletAssignmentComponent deleted.

Receive: compact mobile-first design at every width — QR + latest virgin
address + copy rows + ONE generate-new button (hidden for single-address
wallets); the past-addresses selector and its Core-only enumeration plumbing
are dropped (address reuse is an anti-pattern; the coins page lists history).
Amount/label stay as secondary URI fields.

Send at phone: recipient and amount lose their section titles (the field
labels carry the meaning) and amount + MAX share one row; FEE/OPTIONS keep
their titles. Desktop unchanged.
